NavBar as a separate component, API_URL as a relative path, package as com.netcracker.ncedu.tlt.impl.class_register;
