+++ client/src/services/auth.service.js
@@ -1,7 +1,6 @@
 import axios from 'axios';
 
-// const API_URL = 'https://secure-class-app.herokuapp.com/api/auth/';
-const API_URL = 'http://localhost:8080/api/auth/';
+const API_URL = '/api/auth/';
 
 class AuthService {
   login(user) {
