+++ client/src/services/user.service.js
@@ -1,8 +1,7 @@
 import axios from 'axios';
 import authHeader from './auth-header';
 
-// const API_URL = 'https://secure-class-app.herokuapp.com/api/test/';
-const API_URL = 'http://localhost:8080/api/test/';
+const API_URL = '/api/test/';
 
 class UserService {
   getPublicContent() {
