+++ client/src/views/Billboard.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="container">
+    <header class="jumbotron">
+      <h3>Billboard</h3>
+    </header>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "Billboard"
+}
+</script>
+
+<style scoped>
+
+</style>
