+++ client/src/views/Chat.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="container">
+    <header class="jumbotron">
+      <h3>Chat</h3>
+    </header>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "Chat"
+}
+</script>
+
+<style scoped>
+
+</style>
