+++ client/src/views/FindStudent.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="container">
+    <header class="jumbotron">
+      <h3>FindStudent</h3>
+    </header>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "FindStudent"
+}
+</script>
+
+<style scoped>
+
+</style>
