+++ client/src/views/HomeWork.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="container">
+    <header class="jumbotron">
+      <h3>HomeWork</h3>
+    </header>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "HomeWork"
+}
+</script>
+
+<style scoped>
+
+</style>
