+++ client/src/views/Schedule.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="container">
+    <header class="jumbotron">
+      <h3>Schedule</h3>
+    </header>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "Schedule"
+}
+</script>
+
+<style scoped>
+
+</style>
