+++ server/src/main/java/com/netcracker/ncedu/tlt/impl/class_register/SpringBootSecurityJwtApplication.java
@@ -1,0 +1,13 @@
+package com.netcracker.ncedu.tlt.impl.class_register;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBootSecurityJwtApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringBootSecurityJwtApplication.class, args);
+	}
+
+}
