+++ server/src/main/java/com/netcracker/ncedu/tlt/impl/class_register/models/ERole.java
@@ -1,0 +1,7 @@
+package com.netcracker.ncedu.tlt.impl.class_register.models;
+
+public enum ERole {
+	ROLE_USER,
+    ROLE_MODERATOR,
+    ROLE_ADMIN
+}
