+++ server/src/main/java/com/netcracker/ncedu/tlt/impl/class_register/models/Role.java
@@ -1,0 +1,39 @@
+package com.netcracker.ncedu.tlt.impl.class_register.models;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "roles")
+public class Role {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Integer id;
+
+	@Enumerated(EnumType.STRING)
+	@Column(length = 20)
+	private ERole name;
+
+	public Role() {
+
+	}
+
+	public Role(ERole name) {
+		this.name = name;
+	}
+
+	public Integer getId() {
+		return id;
+	}
+
+	public void setId(Integer id) {
+		this.id = id;
+	}
+
+	public ERole getName() {
+		return name;
+	}
+
+	public void setName(ERole name) {
+		this.name = name;
+	}
+}
