+++ server/src/main/java/com/netcracker/ncedu/tlt/impl/class_register/payload/response/MessageResponse.java
@@ -1,0 +1,17 @@
+package com.netcracker.ncedu.tlt.impl.class_register.payload.response;
+
+public class MessageResponse {
+	private String message;
+
+	public MessageResponse(String message) {
+	    this.message = message;
+	  }
+
+	public String getMessage() {
+		return message;
+	}
+
+	public void setMessage(String message) {
+		this.message = message;
+	}
+}
