+++ server/src/main/java/com/netcracker/ncedu/tlt/impl/class_register/security/services/UserDetailsServiceImpl.java
@@ -1,0 +1,26 @@
+package com.netcracker.ncedu.tlt.impl.class_register.security.services;
+
+import com.netcracker.ncedu.tlt.impl.class_register.models.User;
+import com.netcracker.ncedu.tlt.impl.class_register.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+public class UserDetailsServiceImpl implements UserDetailsService {
+	@Autowired
+    UserRepository userRepository;
+
+	@Override
+	@Transactional
+	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+		User user = userRepository.findByUsername(username)
+				.orElseThrow(() -> new UsernameNotFoundException("User Not Found with username: " + username));
+
+		return UserDetailsImpl.build(user);
+	}
+
+}
