--- server/src/main/java/com/pozhenskii/springjwt/SpringBootSecurityJwtApplication.java
@@ -1,13 +1,0 @@
-package com.pozhenskii.springjwt;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringBootSecurityJwtApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringBootSecurityJwtApplication.class, args);
-	}
-
-}
