--- server/src/main/java/com/pozhenskii/springjwt/models/ERole.java
@@ -1,7 +1,0 @@
-package com.pozhenskii.springjwt.models;
-
-public enum ERole {
-	ROLE_USER,
-    ROLE_MODERATOR,
-    ROLE_ADMIN
-}
