--- server/src/main/java/com/pozhenskii/springjwt/models/Role.java
@@ -1,39 +1,0 @@
-package com.pozhenskii.springjwt.models;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "roles")
-public class Role {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
-
-	@Enumerated(EnumType.STRING)
-	@Column(length = 20)
-	private ERole name;
-
-	public Role() {
-
-	}
-
-	public Role(ERole name) {
-		this.name = name;
-	}
-
-	public Integer getId() {
-		return id;
-	}
-
-	public void setId(Integer id) {
-		this.id = id;
-	}
-
-	public ERole getName() {
-		return name;
-	}
-
-	public void setName(ERole name) {
-		this.name = name;
-	}
-}
