--- server/src/main/java/com/pozhenskii/springjwt/payload/response/JwtResponse.java
@@ -1,74 +1,0 @@
-package com.pozhenskii.springjwt.payload.response;
-
-import java.util.List;
-
-public class JwtResponse {
-	private String token;
-	private String type = "Bearer";
-	private Long id;
-	private String username;
-	private String email;
-	private String photo;
-	private List<String> roles;
-
-	public JwtResponse(String accessToken, Long id, String username, String email, String photo, List<String> roles) {
-		this.token = accessToken;
-		this.id = id;
-		this.username = username;
-		this.email = email;
-		this.photo = photo;
-		this.roles = roles;
-	}
-
-	public String getAccessToken() {
-		return token;
-	}
-
-	public void setAccessToken(String accessToken) {
-		this.token = accessToken;
-	}
-
-	public String getTokenType() {
-		return type;
-	}
-
-	public void setTokenType(String tokenType) {
-		this.type = tokenType;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getPhoto() {
-		return photo;
-	}
-
-	public void setPhoto(String photo) {
-		this.photo = photo;
-	}
-
-	public String getUsername() {
-		return username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	public List<String> getRoles() {
-		return roles;
-	}
-}
