--- server/src/main/java/com/pozhenskii/springjwt/payload/response/MessageResponse.java
@@ -1,17 +1,0 @@
-package com.pozhenskii.springjwt.payload.response;
-
-public class MessageResponse {
-	private String message;
-
-	public MessageResponse(String message) {
-	    this.message = message;
-	  }
-
-	public String getMessage() {
-		return message;
-	}
-
-	public void setMessage(String message) {
-		this.message = message;
-	}
-}
