--- server/src/main/java/com/pozhenskii/springjwt/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.pozhenskii.springjwt.repository;
-
-import java.util.Optional;
-
-import com.pozhenskii.springjwt.models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-	Optional<User> findByUsername(String username);
-
-	Boolean existsByUsername(String username);
-
-	Boolean existsByEmail(String email);
-}
