+++ server/src/main/resources/application.properties
@@ -1,13 +1,13 @@
 
 #heroku
-spring.datasource.url=jdbc:${SPRING_DATASOURCE_URL:postgres://aunxbywnqquojs:73314bff392076498bba4565d4d17c815cc8f6919a91b66fc73bdc438350e673@ec2-46-137-84-140.eu-west-1.compute.amazonaws.com:5432/d1820i16qoc8ed}
-spring.datasource.username=${SPRING_DATASOURCE_USERNAME:aunxbywnqquojs}
-spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:73314bff392076498bba4565d4d17c815cc8f6919a91b66fc73bdc438350e673}
+#spring.datasource.url=jdbc:${SPRING_DATASOURCE_URL:postgres://aunxbywnqquojs:73314bff392076498bba4565d4d17c815cc8f6919a91b66fc73bdc438350e673@ec2-46-137-84-140.eu-west-1.compute.amazonaws.com:5432/d1820i16qoc8ed}
+#spring.datasource.username=${SPRING_DATASOURCE_USERNAME:aunxbywnqquojs}
+#spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:73314bff392076498bba4565d4d17c815cc8f6919a91b66fc73bdc438350e673}
 
 #local
-#spring.datasource.url=jdbc:${SPRING_DATASOURCE_URL:postgresql://localhost:5432/postgres}
-#spring.datasource.username=${SPRING_DATASOURCE_USERNAME:postgres}
-#spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:1234}
+spring.datasource.url=jdbc:${SPRING_DATASOURCE_URL:postgresql://localhost:5432/postgres}
+spring.datasource.username=${SPRING_DATASOURCE_USERNAME:postgres}
+spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:1234}
 
 spring.jpa.hibernate.ddl-auto= update
